♻️ refactor:page 폴더 분리
- page 폴더 분리
- route에 errorPage 적용
Issue Resolved: #163
