+++ front/src/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 
-import Navbar from '@/components/Navbar';
+import Navbar from '@/components/Navbar/index.tsx';
 
 import LoadingPage from '@/pages/LoadingPage';
 
@@ -9,11 +9,11 @@ export default function Layout() {
   return (
     <>
       <Navbar />
-      <Suspense fallback={<LoadingPage />}>
-        <div className="mx-auto my-4 flex h-[calc(100vh-72px)] max-w-[1080px] items-center justify-center">
+      <div className="mx-auto my-4 flex h-[calc(100vh-72px)] max-w-[1080px] items-center justify-center">
+        <Suspense fallback={<LoadingPage />}>
           <Outlet />
-        </div>
-      </Suspense>
+        </Suspense>
+      </div>
     </>
   );
 }
