+++ front/src/routes/index.tsx
@@ -5,12 +5,14 @@ import WithoutLogin from '@/components/loaders/WithoutLogin';
 
 import AdminPage from '@/pages/AdminPage';
 import LoginPage from '@/pages/LoginPage';
+import NotFondPage from '@/pages/NotFoundPage.tsx';
 import ProgramDetailPage from '@/pages/ProgramDetailPage';
 import ProgramsPage from '@/pages/ProgramsPage';
 import ReservationPage from '@/pages/ReservationPage';
 import ReservationWaitingPage from '@/pages/ReservationWaitingPage';
-import SignUpPage from '@/pages/SignUpPage';
+import SignUpPage from '@/pages/SignupPage';
 
+// import Test from '@/pages/Test.tsx';
 import Layout from '@/layout/Layout';
 
 //TODO lazyloading,suspene, fallback 적용, withLogin hoc접근 권한 설정, flat보다는 next 처럼 밑으로 최적화도 더 좋다
@@ -18,9 +20,10 @@ const router = createBrowserRouter([
   {
     path: '/',
     element: <Layout />,
-    errorElement: <div>notFound</div>,
+    errorElement: <NotFondPage />,
     children: [
-      { path: '', element: <Navigate to="programs" /> },
+      { path: '*', element: <NotFondPage /> },
+      { path: '', element: <Navigate to="/programs" /> },
       { path: '/programs', element: <ProgramsPage /> },
       { path: '/programs/:programId', element: <ProgramDetailPage /> },
       {
